TOTALLY NEW VERSION OF FRAMEWORK HEAVILY REVISED AFTER 2 YEARS OF DEVELOPMENT AND USE.
I found many things where I was trying to mimic laravel caused code to be longer and so I went back to the drawing board and make changes to optimize the code and make it easier. This is the first revision of the updates as I merge them into the original AscendPHP framework.